refactor(object-class): rename get-class
