+++ src/lib/index.js
@@ -1,2 +1,2 @@
 export { default as mapKeys } from './map-keys'
-export { default as objectClass } from './object-class'
+export { default as getClass } from './get-class'
